requirements.txt
flask==3.1.3
flask-bcrypt==1.0.1
flask-cors==5.0.1
gunicorn==23.0.0
python-dotenv==1.0.1
psycopg2-binary
cloudinary
requests
